arch: migrate telegram_bot.graph.config + embeddings to src.runtime (#2045 slice 2)
Second slice of the reverse-layering fix tracked under #1948 / #2045 /
#2049. Continues the work from #2096 (graph.state slice 1) by moving
GraphConfig and the BGE-M3 LangChain Embeddings wrappers into the
runtime kernel so src/api/main.py no longer reaches back into
telegram_bot/ for them.

What this PR does:

- Adds src/runtime/integrations/__init__.py and
  src/runtime/integrations/embeddings.py containing the canonical
  BGEM3Embeddings / BGEM3SparseEmbeddings / BGEM3HybridEmbeddings
  implementations copied verbatim from
  telegram_bot/integrations/embeddings.py. The new module imports the
  observability and BGE-M3 client APIs from the existing src.* shims
  (src.observability, src.services.bge_m3_client) instead of the
  telegram_bot.* re-exports, so it satisfies the layering contract.
- Adds src/runtime/graph/config.py containing the canonical GraphConfig
  dataclass + from_env() + create_llm() / create_supervisor_llm() /
  create_embeddings() / create_sparse_embeddings() /
  create_hybrid_embeddings() factory methods. The factories now import
  the embedding wrappers from src.runtime.integrations.embeddings.
- Reduces telegram_bot/integrations/embeddings.py and
  telegram_bot/graph/config.py to thin re-export shims so every existing
  caller (src/api, src/voice, telegram_bot internals, plus 30+ test
  files) keeps working without churn.
- Switches src/api/main.py:37 to import GraphConfig from
  src.runtime.graph.config, removing one of the four remaining reverse
  imports from a hot path.
- Shrinks tests/data/known_layering_violations.json from 4 entries to 3
  for src/api/main.py: telegram_bot.graph.config is now compliant.

What this PR does NOT do (out of scope, follow-up work):

- The remaining 3 src/api/main.py imports (telegram_bot.graph.graph,
  telegram_bot.integrations.cache, telegram_bot.services.qdrant) still
  flow back into telegram_bot. They are tracked under #2047
  (integrations.cache / services.qdrant) and the third graph slice.
  Issue #2049 stays open until the allowlist is empty.
- mini_app separability verification is not part of this slice.

Validation:

- uv run --frozen pytest tests/contract/test_layering_no_telegram_bot_imports_contract.py -q
  → 4 passed (allowlist drift detector picks up the shrunk JSON).
- uv run --frozen pytest tests/unit/graph/ tests/unit/integrations/ tests/unit/agents/test_streaming.py -q
  → 660 passed (back-compat shims work for every existing caller).
- ruff check + ruff format clean.
- pre-commit gitleaks/end-of-file-fixer/etc. clean.

Refs #2049 #2045 #1948
